split ethernet driver
